Initial commit: filled in template with my info
I spotted this type of repo on @ampinsk's profile and got curious why she'd  create a repo w/ her username. Then I looked in the commit history and discovered it was a _special_ repository, so I created my own :D

In this commit, I described myself using the template for getting started: the ideas provided by @github (not sure who the actual folks are but I'd love to know).

Note for the future: I think this is a file that I should review every week (or so), clean up, update, etc. Something like a snapshot of me at a point in time.